Add `Exec` event for kube interactive shells in order to catch the return code
After `SessionStart` event we must catch if the execution was successfull and log it in Audit Log. The session can be denied for several reasons like missing executable, RBAC restriction to the user/group...

The `Exec` will only catch the initial command status (i.e. executable not found, forbidden, exit code) and the session recording will record the user actions in the shell.

Fixes #12283