Improve `kube exec` Audit Log events (#12831)
* Add `session.start` and `session.end` events for non-interactive kube exec

This commit adds the `session.start` and `session.end` events to the non-interactive exec commands.
Before this commit, the `Exec` event was not recorded in the Audit Log because it missed the `Metadata.Code` field and failed the validation.
This code is only known after the exec is triggered (FAILURE or SUCCESS) and so it requires that `Exec` event insertion into the Audit Log must be delayed until the execution finishes (i.e. like in SSH).
Given that, to prevent delayed log entries (execution is blocked for a long time), we require the usage of  `SessionStart` and `SessionEnd` events even if the session is non-interactive.

Fixes #12764

* Add `Exec` event for kube interactive shells in order to catch the return code

After `SessionStart` event we must catch if the execution was successful and log it in the Audit Log. The session can be denied for several reasons like missing executable, RBAC restriction to the user/group...

The `Exec` will only catch the initial command status (i.e. executable not found, forbidden, exit code) and the session recording will record the user actions in the shell.

Fixes #12283