Fix dialing nonexistent auth servers
Provide the list of auth servers from configuration to the reverse
tunnel subsystem so that they can be used to establish connections
to the remote auth server instead of fetching individual auth server
addresses from the backend. This prevents any expired auth servers
from getting into the list to be dialed.