Make `TestDefaultTemplateRendering` less failure prone
The test requires the Teleport server to report a sane public proxy
address, however occasionally it takes some time to initialize and
returns an invalid value until that point.

This adds a wait loop to the client initialization helper to wait up
to 10 seconds for the server to return a sane value, which should
improve reliability of the test.