[v9] Fix broken version check in tbot's `tshwrap` (#13034) (#13037)
* Fix broken version check in tbot's `tshwrap`

`tshwrap` performs a tsh version check to ensure it has the
functionality we need. Unfortunately, during a final refactoring
before merging, we changed the function signature of `capture()` to
require an explicit path to a `tsh` binary but in a way that was
unfortunately not caught by the compiler. The previous syntax meant
we just tried to execute the first argument, i.e. `version`, which
is never what we want.

This PR correctly passes the tsh path to `capture()` to fix the
version check.

* Release 9.3.2

Co-authored-by: Roman Tkachenko <roman@goteleport.com>