Flesh out CAP instructions (#12544)
Closes #11840

Since Cloud accounts begin with a cluster_auth_preference resource,
you need to obtain your current resource via tctl get and make changes,
rather than creating a fresh one. This changes Cloud instructions in
several guides to reflect this.

Also use the same instructions for self-hosted users. If a CAP does
not exist on the backend, the shell redirection used in the
"tctl get" command will result in an empty file, which follows the
existing instructions with minimal changes.

Also update the instructions related to u2f in the Reducing the
Blast Radius guide.