Match full magic byte sequences with detectProto (#12426)
Previously, 8 bytes was always passed into 'detectProto'. But assumptions
were made that the input was only 3 bytes, causing 'detectProto' to match
against inputs that were prefixed with a part of a magic byte sequence.
Magic byte sequences by definition need to be fully present in order to be
effective, so don't match against truncated magic byte sequences, and
do an extra read if necessary to match the ProxyV2 protocol. The length of
the magic bytes of that protocol is 12 bytes, but many other protocols will
only send 8 bytes at first, so only do an extra read if we're reasonably sure
the input is using ProxyV2 to avoid EOF errors.