Add a link from the older docs versions page (#13074)
Backports #12838

If a reader uses the docs version picker to select "Older Versions",
they will navigate to a page where the current version is still
the version they had previously selected. This change adds a link to
the main docs site to make navigation easier.

This is a provisional solution to tide us over until we have a better
way to handle unsupported versions in our docs version picker.