Reorganize the docs homepage menu
Currently, the main body content of the docs home page links to
sections related to individual resources (Server Access, Application
Access, etc.). For users visiting the docs for the first time, it's
difficult to determine what is involved in getting started with
Teleport.

This change organizes the docs landing page to imply that there is a
progression from one stage of the user's setup to the next.

See #12787

- Add headings for different stages of setting up Teleport.
- Add links to a Getting Started guide for each edition that includes a
  "scope" query so users are given the appropriate scope (this partially
  addresses #12773).
- Edit the initial list of Teleport benefits to be more general and
  encompass more functionality.
- Very light copy-editing of the text in the tile lists at the bottom
  of the page.