Ignore Conditional Checks for Existing Items
We use a Conditional Check in DynamoDB for creates to ensure that an
item (specifically, the FullPath) doesn't exist ("attribute_not_exists")

This fails during startup for the following paths:
- teleport/cluster_configuration/name
- teleport/roles/editor/params
- teleport/roles/access/params
- teleport/roles/auditor/params

and periodically (every few minutes) for the following path:
- teleport/cloud/lock

Because these items already exist, and don't need to be created.

This doesn't apply when the mode is to update. This is especially
important because these errors are impacting the
"backend_write_requests_failed_total" metric, which will throw off our
SLO calculations.

This change ignores such errors, returning nil.

Signed-off-by: Michael McAllister <michael.mcallister@goteleport.com>