Backport #12682 to branch/v9 (#12950)
* Add role notes to the teleport-kube-agent intro

Fixes #10480

Previously, the "Connect a Kubernetes Cluster to Teleport" guide
instructed the user to run "tsh kube login" to a cluster they
may not be authorized to access. This change adds explicit instructions
to authorize your user's Teleport role to interact with your cluster.

* Respond to PR feedback

Use the "view" group instead of "system:masters" in the example role.