Add Role and User reconciler
This commit adds a reconciler for the Role and User Resources
Everytime we receive an event for one of those resources, we must propagate that
change into the Teleport cluster.

We have two types of events: update/create and delete

For creating/updating we check if the resource exists in Teleport and
upsert the entire resource based on the current state presented on the
k8s side.

For deleting, the recommendation is to use finalizers
Finalizers allow us to map an external resource to a K8S resource.
So, when we create or update a resource, we add our own finalizer to the K8S
resource list of finalizers
When we receive a delete event, we check if the finalizer that we added is
present, if so, we delete the resource in Teleport and remove the finalizer
We do no-op if the finalizer is not there
K8S will delete the resource when the list of finalizers is empty