Add Sidecar
We are going to use a sidecar architecture to run the Operator.

It will have privileged access because it runs right next to the
Teleport Auth Server, sharing the data dir (`/var/lib/teleport`) and its
configuration (`/etc/teleport/teleport.yaml`).

This will allow us to create the necessary user and role, and an
identity file which the operator will use to create a teleport client
and manage the cluster.