Deflake TestX11Forward
This test fails similarly to TestAgentForward, because the
file-based session uploader runs in the background and is
sometimes still writing to disk when the test goes to clean
up temp directories.

Since this test has nothing to do with the disk-based uploader
or the audit log, we use a discard emitter and set the session
recording mode to synchronous so there are no files on disk
which need to be uploaded.