This MR introduces Kubernetes Secret storage for Teleport Kubernetes Agent state.
When running Teleport Agent inside a Kubernetes cluster where TELEPORT_REPLICA_NAME and KUBE_NAMESPACE exist, the agent will store identities and states in a Kubernetes Secret instead of local storage.

Otherwise, if the above environment variables are not present or not in Kubernetes, the agent persists state in SQLite.

Notes:

- Fresh installs will ignore the storage settings and enable the Kubernetes Secret Storage by default.
- If the operator is upgrading from an existing release that didn't support Secret storage, the Helm chart will continue to use the same object (Statefulset if storage was enabled, otherwise Deployment) until Teleport 11 where we force the migration to Kube Secret Storage.

Implementation of #12958
Related to #5585