Extract fetchRoleSet from tsh/db to lib/client
gravitational/teleport#12281 fixed fetching db users for remote clusters
for `tsh db ls`. gravitational/teleport#13617 applied the same fix to
`tsh db ls --all` and extracted the `fetchRoleSet` function.

This commit extracts it to `lib/client` so that we can reuse it in
`lib/teleterm`. Other than accepting `log` as an argument, nothing was
changed about that function.