Fix CHANGELOG links for new resolution method
In gravitational/docs, relative links in partials are revaluated
relative to the filepath of the partial, not to the page that includes
that partial. In other words, a link in the same partial will be
resolved differently depending on the filepath of the page that
includes it.

To accommdate this link resolution method, this change edits relative
links within the CHANGELOG to be relative to that file, rather than to
docs/pages/changelog.mdx, which includes it.