Reduce use of centralised port allocation in tests
Ports used by the unit tests have been allocated by pulling them out of
a list, with no guarantee that the port is not actually in use. This
central allocation point means that tests cannot be split into separate
packages to be run in parallel, as the ports allocated between the
various packages will be allocated multiple times  and end up
intermittently clashing.

There is also no guarantee, even when the tests are run serially, that
the ports will not clash with services already running on the machine.

This patch (largely) replaces the use of this centralised port
allocation with pre-created listeners injected into the test via the
file descriptor import mechanism use by Teleport to pass open ports to
child processes.