Fix tracing exporter endpoints
Ensure that the endpoint provided to the trace clients
are correct even if the configuration doesn't include
the scheme. Prior to this the endpoint always attempted
to remove the scheme prefix, even when one wasn't provided.
Doing so led to the hostname to be altered which caused
some unknown host issues.

This also removes the process and process owner detector
from the tracing resource. Running within a container might
not have a username mapped to the uid which was preventing
tracing from being initialized.