Improve checkboxes/crosses in comparison table
Small change, but I think the sharper corners on checkboxes look a bit more professional than the rounded ones.  I also added matching crosses with the same weight for when the feature doesn't exist in an edition.

Tested this in the latest Chromium/Brave/Safari/Edge and all symbols show up.  It's possible they won't show in old Internet Explorer versions (like version 5), but I don't have a machine with that to confirm with.  I doubt we get many IE5 visitors though.