Only install cbindgen if not present (#485)
The `cbindgen` binary is now expected to be provided by the build
environment (see gravitational/teleport#14036). Installing over the
top of it with `cargo install cbindgen` now causes an error as cargo
refuses to install over an arbitrary binary with no associated build
metadata.

This patch updates the Makefile to only install `cbindgen` if it is not
already installed.

Source-commit: gravitational/teleport.e@b5d904835ad2c095623c26bed14e660e2eaec68d
Source-pull-requests: gravitational/teleport.e#485