[v10] Retry UV failures and correctly handle single accounts (#14312)
Address the following FIDO2 issues:

1. Retry assertions without UV if UV=true if the PIN is set 2. Correctly handle
single account responses from authenticators

1 happens in the wild and the fix is confirmed. The fix is based in the fact
that the authenticator will set the UV bit in the resulting assertion when the
PIN is present, regardless of us asking for it. We do need the UV bit to clear
server-side validation, that haven't changed.

2 was a misunderstanding on my part; the spec doesn't require authenticators to
return user names for single account responses, which they effectively don't. In
practical terms we are unable to guarantee the `--user` flag for single account
scenarios, but not failing seems better than failing. See the
[authenticatorGetAssertion response][1], user member.

[1]: https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-20210615.html#authenticatorgetassertion-response-structure

Backport #14198 to branch/v10

* Add tests for UV failure and single account
* Correctly handle single account responses
* Retry without UV if a PIN is set