Improve cluster ID is required error
This slightly improves the cluster ID is required error that commonly occurs when you're using a newer version of the tsh client than the cluster supports.  For example, a v8 tsh client with a v7 cluster.  

This error may also occur in other situations, for example, if you followed the wrong upgrade path when upgrading the cluster.  (upgrading from v6 to v8 directly)

This error has caused confusion with customers and employees (using both Cloud and Core) for a good while now and I'd like to improve it.  Any further changes to word this better / make it more accurate are welcome.