Fix makeClientForProxy user extraction
When attempting to connect to a host with an explicit username,
i.e `tsh ssh alice@host`, `makeClientForProxy` will attempt to detect
the username if one is provided. In the case a user is found the
CLI argument provided by the user is updated to removed the user.
This results in any subsequent calls to `makeClientForProxy` to
miss out on the explicit user and attempt to connect as the currently
logged in Teleport user.

To recitfy this `makeClientForProxy` first assigns `cf.UserHost` to
a local variable, and that local variable is then the one that gets
the user removed from it. This allows multiple calls to
`makeClientForProxy` to be able to correctly retrieve the user from
the provided CLI argument.