[v10] Correct Node/agent naming and usage (#14650)
Correct Node/agent naming and usage

Backports #13840

Fixes #10976

This change helps to remedy two issues with the documentation:

- Much of the documentation was written when Teleport's services
  consisted of the Auth Service, Proxy Service, and Nodes/SSH Service.
  There are some places in the docs where a reader can mistake the
  Node Service to mean "any resource service."

- Users often associate the term "agent" with a daemon that must run on
  every host in their infrastructure, and usage of this term can make
  Teleport seem more resource intensive than it actually is.

This change corrects mentions of "Node" that stand in for "any service
that access resources in your infrastructure" to account for Teleport's
other resource services. It also replaces mentions of the term "agent"
with "instance," the name of a resource service, or simply "Teleport."

Also edits the "agentless" question in the FAQ to include more detailed
information about each of Teleport's resource services.

This change also does some light copy-editing to correct the
capitalization of service names. It doesn't aim to be comprehensive,
though, since the main purpose of this change is to fix potentially
misleading mentions of "Node" and "agents".

Finally, this change deletes the production.mdx guide, as there is
no way to access this guide due to redirects. This guide includes
some outdated usage of "Nodes".