[v9] Refactor reversetunnel localsite  (#14792)
Refactor reversetunnel localsite (#14724)

* Refactor reversetunnel localsite

Two major improvements made:

1) Replace []*localsite of the `reversetunnel.Server` with
   *localsite since there will only ever be one of them per Proxy.
2) Reuse the Proxy cache within the localsite instead of creating
   a new cache that is a **subset** of the Proxy cache.

The `reversetunnel.Server`, which is only run on the Proxy, creates
a localsite to track any local connections when it is created. That
localsite is the only one that is ever created, yet the
`reversetunnel.Server` maintained a slice of them. In order to reduce
complexity, we can remove the slice and only track the single site
created with the `reversetunnel.Server`.

When a localsite is created it instantiates a new cache with target
`remote-proxy` which is used to satisfy the RemoteSite interface.
However, `remote-proxy` is a subset of the cache already created
when intializing the Proxy. Which means that a Proxy has **TWO**
caches that monitor the **SAME** set of resources. By simply reusing
the proxy cache when creating the localsite we can reduce the
network bandwidth consumed by an individual Proxy and reduce the
load on the Auth server.