useDocumentGateway: Default to '' rather than '0' (#1061) (#1063)
If '0' is provided as the default and the initial call to create a gateway
fails, the user will be presented with an input that says 0. 0 is an invalid
value though that this input will not accept.

An empty string, on the other hand, is a perfectly valid value (the input
doesn't have the `required` attribute). In that scenario, the backend will
default the value to '0' and pick a random port.