[v8] Fix makeClient user extraction (#14850)  (#14941)
Fix makeClient user extraction

When attempting to connect to a host with an explicit username,
i.e `tsh ssh alice@host`, `makeClient` will attempt to detect
the username if one is provided. In the case a user is found the
CLI argument provided by the user is updated to removed the user.
This results in any subsequent calls to `makeClient` to
miss out on the explicit user and attempt to connect as the currently
logged in Teleport user.

To recitfy this `makeClient` first assigns `cf.UserHost` to
a local variable, and that local variable is then the one that gets
the user removed from it. This allows multiple calls to
`makeClient` to be able to correctly retrieve the user from
the provided CLI argument.