[v8] add backport workflow
The backport workflow is now also responsible for deleting the
remote branch after a backport is merged. In order for it to run,
the workflow itself must exist on the branch.

Note: the code doesn't need to be backported, as the workflow always
checks out and runs the code from master.