desktop access: try using system DNS resolver first
Prior to this change, we assumed that the best way to resolve
auto-discovered desktops would be to make a DNS query of the
configured LDAP server, since the Linux box running Teleport
would likely not be domain joined and may have separate DNS
configuration.

It turns out to be pretty common to deploy dedicated LDAP servers
behind a VIP or other frontend, in which case we fail to resolve
the desktop's IP even though the system resolver would have worked
just fine.

Instead, make a standard DNS query using the system's default
settings first, and fall back to the LDAP server only if that
query fails.

Fixes #15154