[v8] Ensure that the WindowsDesktopReady event is emitted (#14859)
Ensure that the WindowsDesktopReady event is emitted

When desktop access is enabled, the TeleportReady event will not
be emitted until the WindowsDesktopReadyEvent is emitted, and it
turns out we have *never* emitted a WindowsDesktopReadyEvent.

This is likely due to desktop access being copied from kube access
since the very beginning. The same issue was recently fixed for
kube access in #9418.