Don't try to emit an end event for an empty session
In some cases, an auth server can see a response from the uploader
with an empty session ID. If this is the case, skip emitting the
session end and upload events, as there's no action to take.

This scenario can happen when multiple auth servers are running
upload completers and working on the same uploads simultaneously.