Fix table formatting in the SOC 2 guide
Some rows of the table within the guide include separate table columns
for different Teleport features that support a single SOC 2 requirement.
The table applies this format inconsistently, causing some table cells
to be hidden.

This change limits each table row to three columns--the number of
columns defined in the first row of the table--and uses line breaks to
separate multiple features that support a single SOC 2 requirement.