Rewrite get-webapps-version in bash
The original thought was that Go would be easier to test and reason
about, but due to the our drone setup works, requiring Go/Docker at
this step of the process introduced extra complexity.

This script is pretty simple and will be removed later this quarter
when we migrate webapps into the Teleport repo.